Move subscription routes into folder. Add errors.ts, add errors to all routes and update tests that check for errors.
