+++ server/router.ts
@@ -35,13 +35,13 @@ import deleteCommunity from './routes/deleteCommunity';
 import updateCommunity from './routes/updateCommunity';
 import viewCount from './routes/viewCount';
 
-import viewSubscriptions from './routes/viewSubscriptions';
-import createSubscription from './routes/createSubscription';
-import deleteSubscription from './routes/deleteSubscription';
-import enableSubscriptions from './routes/enableSubscriptions';
-import disableSubscriptions from './routes/disableSubscriptions';
-import enableImmediateEmails from './routes/enableImmediateEmails';
-import disableImmediateEmails from './routes/disableImmediateEmails';
+import viewSubscriptions from './routes/subscription/viewSubscriptions';
+import createSubscription from './routes/subscription/createSubscription';
+import deleteSubscription from './routes/subscription/deleteSubscription';
+import enableSubscriptions from './routes/subscription/enableSubscriptions';
+import disableSubscriptions from './routes/subscription/disableSubscriptions';
+import enableImmediateEmails from './routes/subscription/enableImmediateEmails';
+import disableImmediateEmails from './routes/subscription/disableImmediateEmails';
 import viewNotifications from './routes/viewNotifications';
 import markNotificationsRead from './routes/markNotificationsRead';
 import clearReadNotifications from './routes/clearReadNotifications';
