+++ server/routes/subscription/disableImmediateEmails.ts
@@ -1,15 +1,16 @@
 import Sequelize from 'sequelize';
 const Op = Sequelize.Op;
 import { Request, Response, NextFunction } from 'express';
-import { factory, formatFilename } from '../util/logging';
+import { factory, formatFilename } from '../../util/logging';
+import Errors from './errors';
 const log = factory.getLogger(formatFilename(__filename));
 
 export default async (models, req: Request, res: Response, next: NextFunction) => {
   if (!req.user) {
-    return next(new Error('Not logged in'));
+    return next(new Error(Errors.NotLoggedIn));
   }
   if (!req.body['subscription_ids[]']) {
-    return next(new Error('must specifiy subscription ids'));
+    return next(new Error(Errors.NoSubscriptionId));
   }
 
   let idOptions;
@@ -24,7 +25,7 @@ export default async (models, req: Request, res: Response, next: NextFunction) =
   });
 
   if (subscriptions.find((s) => s.subscriber_id !== req.user.id)) {
-    return next(new Error('subscription id is not owned by user'));
+    return next(new Error(Errors.NotUsersSubscription));
   }
 
   // TODO: transactionalize this
