+++ server/routes/subscription/errors.ts
@@ -1,0 +1,11 @@
+const Errors = {
+  NotAdmin: 'Not an admin of chain or community',
+  NotLoggedIn: 'Not logged in',
+  NoCategoryAndObjectId: 'Must provide category and object id',
+  InvalidNotificationCategory: 'invalid notification category',
+  NoSubscriptionId: 'Must provide subscription id(s)',
+  NoSubscription: 'Subscription not found',
+  NotUsersSubscription: 'Subscription does not belong to user',
+};
+
+export default Errors;
