+++ server/routes/subscription/viewSubscriptions.ts
@@ -1,10 +1,11 @@
 import { Request, Response, NextFunction } from 'express';
-import { factory, formatFilename } from '../util/logging';
+import { factory, formatFilename } from '../../util/logging';
+import Errors from './errors';
 const log = factory.getLogger(formatFilename(__filename));
 
 export default async (models, req: Request, res: Response, next: NextFunction) => {
   if (!req.user) {
-    return next(new Error('Not logged in'));
+    return next(new Error(Errors.NotLoggedIn));
   }
 
   const subscriptions = await models.Subscription.findAll({
